requirements.txt
# python==3.9.20
# huggingface-hub==0.26.3
sentence_transformers==3.3.1
Flask
nltk
unstructured
loguru
spark_ai_python
langchain-text-splitters
langchain-community==0.3.9
gradio
faiss-cpu
# langchain
